+++ packages/auth/CHANGELOG.md
@@ -1,3 +1,10 @@
+## @lunora/auth [1.0.0-alpha.144](https://github.com/anolilab/lunora/compare/@lunora/auth@1.0.0-alpha.143...@lunora/auth@1.0.0-alpha.144) (2026-09-19)
+
+
+### Dependencies
+
+* **@lunora/server:** upgraded to 1.0.0-alpha.130
+
 ## @lunora/auth [1.0.0-alpha.143](https://github.com/anolilab/lunora/compare/@lunora/auth@1.0.0-alpha.142...@lunora/auth@1.0.0-alpha.143) (2026-09-13)